convert double to long value through call to .longValue()
git-svn-id: http://pubsvn.inavare.net/openscada/modules/atlantis/trunk@4219 1b50e8ee-deed-0310-b1de-ab401e1c4d63

+++ openscada_core_common/src/org/openscada/core/Variant.java
@@ -315,15 +315,15 @@ public class Variant
         {
             if ( this.value instanceof Boolean )
             {
-                return ( (Boolean)this.value ).booleanValue () ? 1 : 0;
+                return ( (Boolean)this.value ).booleanValue () ? 1L : 0L;
             }
             if ( this.value instanceof Double )
             {
-                return ( (Double)this.value ).intValue ();
+                return ( (Double)this.value ).longValue ();
             }
             if ( this.value instanceof Integer )
             {
-                return ( (Integer)this.value ).intValue ();
+                return ( (Integer)this.value ).longValue ();
             }
             if ( this.value instanceof Long )
             {
